Fix: yarn build 되게 오류 수정
